Refactor the index.js to route to either login/signup pages or the main view
Keep track of login status inside of the state in index.js, update the state within the login or signup page submission on success methods.
Successful submission of login/signup will reroute the user to the main view, in which they see tabs and can manually route the view themselves.
Need to add navigation to these tab views.
